chore(playground): refactor to use useLoop and sub components
